# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "مستندات API سامانه لد ",
        "description": "توضیحات نحوه استفاده از APIها",
        "contact": {
            "email": "hreza779@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "سرور اصلی API"
        }
    ],
    "paths": {
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get user's notifications",
                "operationId": "911565ad9aff9ba7eded3f189509e578",
                "parameters": [
                    {
                        "name": "is_read",
                        "in": "query",
                        "description": "Filter by read status (1 or 0)",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of notifications"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Send a manual notification (Admin only)",
                "operationId": "91f8e779a9e46cb5cdda969e5a928a94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "title",
                                    "message"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Welcome!"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Welcome to our platform."
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "system"
                                    },
                                    "action_url": {
                                        "type": "string",
                                        "example": "/dashboard"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notification sent successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/{notification}/read": {
            "put": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark notification as read",
                "operationId": "679a5888e7bc3332a6add3ebc0eafc7a",
                "parameters": [
                    {
                        "name": "notification",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification marked as read"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/mark-all-read": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark all notifications as read",
                "operationId": "e1e017838e0e1ed1cd321ccf80869335",
                "responses": {
                    "200": {
                        "description": "All notifications marked as read"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/{notification}": {
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Delete notification",
                "operationId": "7f078e4ee294fe830b4b842939b15a12",
                "parameters": [
                    {
                        "name": "notification",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Notification deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}